Clean cmd package by applying Standard Go Project Layout
To improve the readability of the repository for new contributors, we
introduce Standard Go Project Layout.
